+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Deprecated `Model::totalRecords()`.
 - When the iterator gets a model count it reuses the same query instead of creating a new one.
 - The deprecated `relation()` method now performs a lookup on the data layer and can return null values.
+- `Model::find()` now uses `queryModels()` on the data driver instead of using `loadModel()`.
 
 ## 0.2.0 - 2017-05-29
 ### Added